backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
pymongo==4.8.0
llama-cpp-python==0.2.90
ultralytics==8.2.82
insightface==0.7.3
opencv-python==4.10.0.84
numpy==1.26.4
python-socketio==5.11.4
eventlet==0.36.1
requests==2.31.0
onnxruntime==1.17.0
onnx==1.15.0
gunicorn==21.2.0
PyJWT==2.8.0
